requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
chromadb>=0.4.22
tiktoken>=0.5.0
fastapi>=0.100.0
uvicorn>=0.20.0
